+++ src/main/java/com/formmaker/fff/answer/AnswerRepository.java
@@ -1,4 +1,4 @@
-package com.formmaker.fff.survey;
+package com.formmaker.fff.answer;
 
 import com.formmaker.fff.answer.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
