+++ src/main/java/com/formmaker/fff/survey/SurveyService.java
@@ -2,15 +2,12 @@ package com.formmaker.fff.survey;
 
 
 import com.formmaker.fff.answer.Answer;
-import com.formmaker.fff.answer.AnswerCreateRequestDto;
+import com.formmaker.fff.answer.AnswerRepository;
 import com.formmaker.fff.question.Question;
-import com.formmaker.fff.question.QuestionCreateRequestDto;
+import com.formmaker.fff.question.QuestionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
